Refactor convo_parser and add DRC and LVS integration for convo generation
