+++ openfasoc/generators/glayout/glayout/llm/convo_parser/Command.py
@@ -1,3 +1,5 @@
+from parse_utils import parseKeyValues, parseKwarg, parseTuple, regenKeyValues, regenKwargs, regenTuple
+
 class Command:
     def __init__(self, type: str):
         self.type = type
@@ -142,124 +144,10 @@ class Comment(Command):
         self.type = "comment"
     def regenCommand(self):
         return self.line
-
-
-def parseKeyValues(line: str):
-    key = ""
-    value = ""
-    keyToken = True
-    keyValue = dict()
-    kwargStart = False
-    tupleStart = False
-    for ch in line:
-        if ch == "{":
-            kwargStart = True
-        elif ch == "(":
-            tupleStart = True
-        elif ch == "}":
-            kwargStart = False
-            value = parseKwarg(value)
-        elif ch == ")":
-            tupleStart = False
-            value = parseTuple(value)
-        elif ch == "=":
-            keyToken = False
-        elif (ch == " " or ch == ",") and (not kwargStart and not tupleStart):
-            if key != "" and value != "":
-                keyValue[key] = value
-            keyToken = True
-            key = ""
-            value = ""
-        else:
-            if keyToken:
-                key = key + ch
-            else:
-                value = value + ch
-    keyValue[key]= value
-    return keyValue
-
-def parseKwarg(line: str):
-    trimmed = line.replace(' ','')
-    words = trimmed.split(",")
-    kwargs = dict()
-    for word in words:
-        keyVal = word.split(":")
-        kwargs[keyVal[0][1:-1]] = keyVal[1]
-    return kwargs
-
-def parseTuple(line: str):
-    trimmed = line.replace(' ','')
-    words = trimmed.split(",")
-    return (words[0], words[1])
-
-def regenKeyValues(keyValues: dict):
-    keyVals = ""
-    for key in keyValues:
-        if key == "kwargs":
-            value = regenKwargs(keyValues[key])
-        elif type(keyValues[key]) == tuple:
-            value = regenTuple(keyValues[key])
-        else:
-            value = keyValues[key]
-        keyVals += f"{key}={value}, "
-    keyVals = keyVals[:-2]
-    return keyVals
-
-def regenKwargs(kwargs: dict):
-    kwarg = "{"
-    for arg in kwargs:
-        kwarg += f"\'{arg}\': {kwargs[arg]}, "
-    kwarg = kwarg[:-2]
-    kwarg += "}"
-    return kwarg
-
-def regenTuple(tuple):
-    return f"({tuple[0]},{tuple[1]})"
-
-class Convo:
-    def __init__(self, compName):
-        self.compName = compName
-        self.commands = []
-    def addCommand(self, command):
-        self.commands.append(command)
-    def changeParam(self, param, newparam):
-        for command in self.commands:
-            command.changeParamDependency(param, newparam)
-    def changeComp(self, comp, newcomp):
-        for command in self.commands:
-            command.changeCompDependency(comp, newcomp)
-    def unparametrize(self, param, value):
-        for index in range(len(self.commands)):
-            if self.commands[index].type == "create":
-                if self.commands[index].name == param:
-                    del self.commands[index]
-                    break
-        self.changeParam(param, value)
-
-    def regen(self):
-        str = self.compName + "\n"
-        for command in self.commands:
-            str += command.regenCommand() + "\n"
-        str = str[:-1]
-        return str
-
-class ConvoParser:
-    def __init__(self, filename: str):
-        self.fileContents = open(filename).read()
-        self.readContents()
-    def readContents(self):
-        commClassMap = {"import":Import, "create":Param, "place":Place, "move": Move, "route":Route}
-        lines = self.fileContents.split('\n')
-        self.convo = Convo(lines[0])
-        for line in lines[1:]:
-            commandType = line.split(" ")[0]
-            if commandType in commClassMap:
-                self.convo.addCommand(commClassMap[commandType](line))
-            elif line[0] == "#":
-                self.convo.addCommand(Comment(line))
-
-convo = ConvoParser("syntax_data/convos/CascodeCommonSourceInterdigitated.convo").convo
-convo.changeComp("CascodeCommonSource", "ccs")
-convo.changeParam("width", "widthCS")
-convo.unparametrize("fingers", "1")
-print(convo.regen())
+    
+class Newline(Command):
+    def __init__(self):
+        self.type = "newline"
+    def regenCommand(self):
+        return ""
+
