+++ openfasoc/generators/glayout/glayout/llm/parser_test.py
@@ -1,0 +1,7 @@
+import sys,os
+sys.path.append('convo_parser')
+
+from convo_parser.convoparser import Convo, ConvoParser
+
+convo = ConvoParser("syntax_data/convos/CurrentMirrorNtype.convo").convo
+print(convo.regen())
